client/requirements.txt
colorzero==2.0
gpiozero==2.0.1
RPi.GPIO==0.7.1
smbus2==0.5.0
websockets==14.1
